requirements.txt
psutil>=5.9.0
pynput>=1.7.6
pywin32>=305
colorama>=0.4.6
pystray>=0.19.4
Pillow>=10.0.0
customtkinter>=5.0.0

PySide6
